Updating code

+++ src/main/java/com/epam/rd/autocode/spring/project/service/impl/BookServiceImpl.java
@@ -1,16 +1,23 @@
 package com.epam.rd.autocode.spring.project.service.impl;
 
 import com.epam.rd.autocode.spring.project.dto.BookDTO;
+import com.epam.rd.autocode.spring.project.exception.AlreadyExistException;
+import com.epam.rd.autocode.spring.project.exception.NotFoundException;
 import com.epam.rd.autocode.spring.project.model.Book;
 import com.epam.rd.autocode.spring.project.repo.BookRepository;
 import com.epam.rd.autocode.spring.project.service.BookService;
+import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.rmi.AlreadyBoundException;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
+@Transactional
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
     private final ModelMapper modelMapper;
@@ -21,7 +28,9 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<BookDTO> getAllBooks() {
+        log.debug("Fetching all books");
         return bookRepository.findAll()
                 .stream()
                 .map(book -> modelMapper.map(book, BookDTO.class))
@@ -29,30 +38,50 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public BookDTO getBookByName(String name) {
+        log.debug("Fetching book by name: {}", name);
         Book book = bookRepository.findByName(name)
-                .orElseThrow(() -> new RuntimeException("Book not found"));
+                .orElseThrow(() -> new AlreadyExistException("Book not found with name: " + name));
         return modelMapper.map(book, BookDTO.class);
     }
 
     @Override
     public BookDTO updateBookByName(String name, BookDTO bookDTO) {
+        log.debug("Updating book with name: {}", name);
         Book book = bookRepository.findByName(name)
-                .orElseThrow(() -> new RuntimeException("Book not found"));
+                .orElseThrow(() -> new NotFoundException("Book not found with name: " + name));
+
+        if (!name.equals(bookDTO.getName()) && bookRepository.existsByName(bookDTO.getName())) {
+            throw new AlreadyBoundException("Book already exists with name: " + bookDTO.getName());
+        }
+
         modelMapper.map(bookDTO, book);
-        bookRepository.save(book);
+        book = bookRepository.save(book);
+        log.info("Book updated successfully: {}", name);
         return modelMapper.map(book, BookDTO.class);
     }
 
     @Override
     public void deleteBookByName(String name) {
+        log.debug("Deleting book with name: {}", name);
+        if (!bookRepository.existsByName(name)) {
+            throw new NotFoundException("Book not found with name: " + name);
+        }
         bookRepository.deleteByName(name);
+        log.info("Book deleted successfully: {}", name);
     }
 
     @Override
     public BookDTO addBook(BookDTO bookDTO) {
+        log.debug("Adding new book: {}", bookDTO.getName());
+        if (bookRepository.existsByName(bookDTO.getName())) {
+            throw new AlreadyExistException("Book already exists with name: " + bookDTO.getName());
+        }
+
         Book book = modelMapper.map(bookDTO, Book.class);
-        bookRepository.save(book);
+        book = bookRepository.save(book);
+        log.info("Book added successfully: {}", book.getName());
         return modelMapper.map(book, BookDTO.class);
     }
 }
